undo class hierarchy changes
